no click true and components modularization
