+++ src/components/Chat/ActiveChat/Messages/components/DropImage/components/ShowImageModal.tsx
@@ -1,8 +1,5 @@
 import React from 'react';
 import {
-  Box,
-  Text,
-  Flex,
   Modal,
   ModalOverlay,
   ModalContent,
